# api/openapi.yaml
openapi: 3.0.3
info:
  title: Walwatch Auto-Renewal Keeper API
  description: API for managing blob auto-renewal on Walrus/Sui.
  version: 0.1.0
  contact:
    name: Walwatch
    url: https://walwatch.io
servers:
  - url: https://api.walwatch.io/api/v1
    description: Production
  - url: http://localhost:3001/api/v1
    description: Local development
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  parameters:
    OrgIdHeader:
      name: X-Org-Id
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Organization UUID for scoping requests
...
  schemas:
...
    SpendingLimit:
      type: object
      properties:
        id:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        amount:
          type: number
        period:
          type: string
        spent:
          type: number
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSpendingLimitInput:
      type: object
      required:
        - walletId
        - amount
      properties:
        walletId:
          type: string
          format: uuid
        amount:
          type: number
        name:
          type: string
        period:
          type: string

    # ── Renewal Job ──
...
paths:
...
  /spending-limits:
    get:
      tags: [Spending Limits]
      summary: List spending limits
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      responses:
        '200':
          description: List of spending limits
          content:
            application/json:
              schema:
                type: object
                properties:
                  spendingLimits:
                    type: array
                    items:
                      $ref: '#/components/schemas/SpendingLimit'
    post:
      tags: [Spending Limits]
      summary: Create spending limit
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSpendingLimitInput'
      responses:
        '201':
          description: Spending limit created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendingLimit'
  /spending-limits/{id}:
    get:
      tags: [Spending Limits]
      summary: Get spending limit
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Spending limit details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendingLimit'
    patch:
      tags: [Spending Limits]
      summary: Update spending limit
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Spending limit updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpendingLimit'
    delete:
      tags: [Spending Limits]
      summary: Delete spending limit
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Spending limit deleted

  # ── Teams ──
...